Fix round two: reload semantics, run-state predicates, and session teardown
The first round's two fixes combined into a regression. Teaching the set path to
record a bank pin, and giving the reload hooks callers, meant a breakpoint was
now re-armed automatically -- but the retry path hard-coded DEBUG_BANK_ANY and
nothing had ever stored the condition, so after any overlay reload a breakpoint
the user had pinned to bank 3 with "A == 5" came back firing in every bank,
unconditionally. The condition and hit count are now kept alongside the
breakpoint and replayed, which also recovers the pin, and the invalidation drops
the old condition record rather than orphaning it for a later arm to trip over.
Retry now reports the breakpoint verified again; it previously stayed silent on
the grounds that the client already believed it was verified, which stopped being
true once invalidation started telling the client otherwise.

The auto-resume for a headless session could never fire. It tested
showDebugOnRender as "is there a UI", but that is assigned from currentMode on
every debugger poll, so being stopped implied it was set and the condition was
false exactly when it mattered. There is now a flag for what was actually meant:
whether a local interactive debugger was asked for.

Execution control is no longer silently dead when the guest clears
debugger_enabled by writing $9FB0. The previous fix polled the socket in that
state but nothing acted on the result -- breakpoints were never tested and the
run state was ignored, so a client was told the machine had paused while it kept
running. The status hook now runs whenever the server is enabled, which also
puts the socket work back behind its 1-in-1000 throttle; the bare poll it
replaces was a syscall per emulated instruction.

Unload ranges are no longer truncated to 16 bits. A range crossing $FFFF
inverted into one that matched nothing, and one entirely above it collapsed onto
$0000-$FFFF and invalidated every unrelated breakpoint.

The three server-side breakpoint tables no longer delete each other's entries.
They add into one core table that deduplicates by (pc, bank, x16Bank), so a
function breakpoint and a source breakpoint on the same address shared a single
entry and whichever was cleared first disarmed the other while the editor still
showed it set.

pause is answered when the machine is already halted, rather than relying on a
transition that did not happen and leaving the request unacknowledged forever.
configurationDone checks for an existing stop before stopOnEntry, so one halt is
not announced twice under two reasons, and gives up if sending the response
dropped the client rather than halting a machine with nobody left to resume it.

Disconnecting now clears every breakpoint kind the session owned, not just the
source ones; function, instruction and data breakpoints were surviving the
client that set them.

An oversized frame drops the client instead of trying to skip bytes that may not
have arrived, which desynchronised the framing and ended in the same disconnect
by a longer route.

+++ src/dbg_load.c
@@ -390,8 +390,7 @@ dbg_load_poll(void)
 		// A DAP client's breakpoints are addresses resolved from debug info
 		// that has just been replaced, so anything inside the range it covered
 		// is now pointing at whatever used to be there.
-		debug_server_invalidate_breakpoints_in_range((uint16_t)new_start,
-		                                            (uint16_t)new_end);
+		debug_server_invalidate_breakpoints_in_range(new_start, new_end);
 
 		// Forgotten whether or not that succeeded. It unloads the range before
 		// reading the file, so a failure part-way through still leaves those
